Remove more warnings

+++ modules/gaia/tasks/readfilejob.cpp
@@ -203,9 +203,15 @@ void ReadFileJob::execute() {
             float radVelZ = 1000.f * radial_vel[i] * rGal.z;
 
             // Use Pythagoras theorem for the final Space Velocity [m/s].
-            values[idx++] = sqrt(pow(radVelX, 2) + pow(tanVelX, 2)); // Vel X [U]
-            values[idx++] = sqrt(pow(radVelY, 2) + pow(tanVelY, 2)); // Vel Y [V]
-            values[idx++] = sqrt(pow(radVelZ, 2) + pow(tanVelZ, 2)); // Vel Z [W]
+            values[idx++] = static_cast<float>(
+                std::sqrt(std::pow(radVelX, 2) + std::pow(tanVelX, 2)) // Vel X [U]
+            );
+            values[idx++] = static_cast<float>(
+                std::sqrt(std::pow(radVelY, 2) + std::pow(tanVelY, 2)) // Vel Y [V]
+            );
+            values[idx++] = static_cast<float>(
+                std::sqrt(std::pow(radVelZ, 2) + std::pow(tanVelZ, 2)) // Vel Z [W]
+            );
         }
         // Otherwise use the vector [m/s] we got from proper motion.
         else {
